WIP: make DQMFileSaver a single file, and save() a function in the module.
